+++ yahoo_market.go
@@ -5,14 +5,13 @@
 package mop
 
 import (
+	"encoding/json"
 	"fmt"
 	"io/ioutil"
 	"net/http"
-	"encoding/json"
 )
 
-// Ongoing issue with Yahoo API version
-const marketURL = `https://query1.finance.yahoo.com/v6/finance/quote?symbols=%s`
+const marketURL = `https://query1.finance.yahoo.com/v7/finance/quote?crumb=%s&symbols=%s`
 const marketURLQueryParts = `&range=1d&interval=5m&indicators=close&includeTimestamps=false&includePrePost=false&corsDomain=finance.yahoo.com&.tsrc=finance`
 
 // Market stores current market information displayed in the top three lines of
@@ -33,6 +32,8 @@ type Market struct {
 	Gold      map[string]string
 	errors    string // Error(s), if any.
 	url       string // URL with symbols to fetch data
+	cookies   string // cookies for auth
+	crumb     string // crumb for the cookies, to be applied as a query param
 }
 
 // Returns new initialized Market struct.
@@ -54,7 +55,9 @@ func NewMarket() *Market {
 	market.Euro = make(map[string]string)
 	market.Gold = make(map[string]string)
 
-	market.url = fmt.Sprintf(marketURL, `^DJI,^IXIC,^GSPC,^N225,^HSI,^FTSE,^GDAXI,^TNX,CL=F,JPY=X,EUR=X,GC=F`) + marketURLQueryParts
+	market.cookies = fetchCookies()
+	market.crumb = fetchCrumb(market.cookies)
+	market.url = fmt.Sprintf(marketURL, market.crumb, `^DJI,^IXIC,^GSPC,^N225,^HSI,^FTSE,^GDAXI,^TNX,CL=F,JPY=X,EUR=X,GC=F`) + marketURLQueryParts
 
 	market.errors = ``
 
@@ -73,7 +76,29 @@ func (market *Market) Fetch() (self *Market) {
 		}
 	}()
 
-	response, err := http.Get(market.url)
+	client := http.Client{}
+	request, err := http.NewRequest("GET", market.url, nil)
+	if err != nil {
+		panic(err)
+	}
+
+	request.Header = http.Header{
+		"Accept":          {"*/*"},
+		"Accept-Language": {"en-US,en;q=0.5"},
+		"Connection":      {"keep-alive"},
+		"Content-Type":    {"application/json"},
+		"Cookie":          {market.cookies},
+		"Host":            {"query2.finance.yahoo.com"},
+		"Origin":          {"https://finance.yahoo.com"},
+		"Referer":         {"https://finance.yahoo.com"},
+		"Sec-Fetch-Dest":  {"empty"},
+		"Sec-Fetch-Mode":  {"cors"},
+		"Sec-Fetch-Site":  {"same-site"},
+		"TE":              {"trailers"},
+		"User-Agent":      {userAgent},
+	}
+
+	response, err := client.Do(request)
 	if err != nil {
 		panic(err)
 	}
@@ -102,39 +127,39 @@ func (market *Market) isMarketOpen(body []byte) []byte {
 
 // -----------------------------------------------------------------------------
 func assign(results []map[string]interface{}, position int, changeAsPercent bool) map[string]string {
-    out := make(map[string]string)
+	out := make(map[string]string)
 	out[`change`] = float2Str(results[position]["regularMarketChange"].(float64))
-        out[`latest`] = float2Str(results[position]["regularMarketPrice"].(float64))
-    if changeAsPercent{
-        out[`change`] = float2Str(results[position]["regularMarketChangePercent"].(float64)) + `%`
-    } else { 
-        out[`percent`] = float2Str(results[position]["regularMarketChangePercent"].(float64))
-    }
-    return out
+	out[`latest`] = float2Str(results[position]["regularMarketPrice"].(float64))
+	if changeAsPercent {
+		out[`change`] = float2Str(results[position]["regularMarketChangePercent"].(float64)) + `%`
+	} else {
+		out[`percent`] = float2Str(results[position]["regularMarketChangePercent"].(float64))
+	}
+	return out
 }
 
 // -----------------------------------------------------------------------------
 func (market *Market) extract(body []byte) *Market {
-    d := map[string]map[string][]map[string]interface{}{}
-    err := json.Unmarshal(body, &d)
-    if err != nil {
-        panic(err)
-    }
-    results := d["quoteResponse"]["result"]
-    market.Dow = assign(results, 0, false)
-    market.Nasdaq = assign(results, 1, false)
-    market.Sp500 = assign(results, 2, false)
-    market.Tokyo = assign(results, 3, false)
-    market.HongKong = assign(results, 4, false)
-    market.London = assign(results, 5, false)
-    market.Frankfurt = assign(results, 6, false)
-    market.Yield[`name`] = `10-year Yield`
-    market.Yield = assign(results, 7, false)
+	d := map[string]map[string][]map[string]interface{}{}
+	err := json.Unmarshal(body, &d)
+	if err != nil {
+		panic(err)
+	}
+	results := d["quoteResponse"]["result"]
+	market.Dow = assign(results, 0, false)
+	market.Nasdaq = assign(results, 1, false)
+	market.Sp500 = assign(results, 2, false)
+	market.Tokyo = assign(results, 3, false)
+	market.HongKong = assign(results, 4, false)
+	market.London = assign(results, 5, false)
+	market.Frankfurt = assign(results, 6, false)
+	market.Yield[`name`] = `10-year Yield`
+	market.Yield = assign(results, 7, false)
 
-    market.Oil = assign(results, 8, true)
-    market.Yen = assign(results, 9, true)
-    market.Euro = assign(results, 10, true)
-    market.Gold = assign(results, 11, true)
+	market.Oil = assign(results, 8, true)
+	market.Yen = assign(results, 9, true)
+	market.Euro = assign(results, 10, true)
+	market.Gold = assign(results, 11, true)
 
-    return market
+	return market
 }
