+++ yahoo_quotes.go
@@ -15,13 +15,7 @@ import (
 	"strings"
 )
 
-// Ongoing issue with Yahoo API version
-// const quotesURLv7 = `https://query1.finance.yahoo.com/v7/finance/quote?symbols=%s`
-// const quotesURL = `https://query1.finance.yahoo.com/v6/finance/quote?symbols=%s`
 const quotesURL = `https://query2.finance.yahoo.com/v7/finance/quote?crumb=%s&symbols=%s`
-const crumbURL = "https://query1.finance.yahoo.com/v1/test/getcrumb"
-const cookieURL = "https://login.yahoo.com"
-const userAgent = "Mozilla/5.0 (Windows NT 10.0; Win64; x64; rv:109.0) Gecko/20100101 Firefox/113.0"
 
 // const quotesURLv7QueryParts = `&range=1d&interval=5m&indicators=close&includeTimestamps=false&includePrePost=false&corsDomain=finance.yahoo.com&.tsrc=finance`
 const quotesURLQueryParts = `&range=1d&interval=5m&indicators=close&includeTimestamps=false&includePrePost=false&corsDomain=finance.yahoo.com&.tsrc=finance`
@@ -72,80 +66,6 @@ func NewQuotes(market *Market, profile *Profile) *Quotes {
 	}
 }
 
-func fetchCrumb(cookies string) string {
-	client := http.Client{}
-	request, err := http.NewRequest("GET", crumbURL, nil)
-	if err != nil {
-		panic(err)
-	}
-
-	request.Header = http.Header{
-		"Accept":          {"*/*"},
-		"Accept-Encoding": {"gzip, deflate, br"},
-		"Accept-Language": {"en-US,en;q=0.5"},
-		"Connection":      {"keep-alive"},
-		"Content-Type":    {"text/plain"},
-		"Cookie":          {cookies},
-		"Host":            {"query1.finance.yahoo.com"},
-		"Sec-Fetch-Dest":  {"empty"},
-		"Sec-Fetch-Mode":  {"cors"},
-		"Sec-Fetch-Site":  {"same-site"},
-		"TE":              {"trailers"},
-		"User-Agent":      {userAgent},
-	}
-
-	response, err := client.Do(request)
-	if err != nil {
-		panic(err)
-	}
-	defer response.Body.Close()
-
-	body, err := ioutil.ReadAll(response.Body)
-	if err != nil {
-		panic(err)
-	}
-
-	return string(body[:])
-}
-
-func fetchCookies() string {
-	client := http.Client{}
-	request, err := http.NewRequest("GET", cookieURL, nil)
-	if err != nil {
-		panic(err)
-	}
-
-	request.Header = http.Header{
-		"Accept":                   {"text/html,application/xhtml+xml,application/xml;q=0.9,image/avif,image/webp,*/*;q=0.8"},
-		"Accept-Encoding":          {"gzip, deflate, br"},
-		"Accept-Language":          {"en-US,en;q=0.5"},
-		"Connection":               {"keep-alive"},
-		"Host":                     {"login.yahoo.com"},
-		"Sec-Fetch-Dest":           {"document"},
-		"Sec-Fetch-Mode":           {"navigate"},
-		"Sec-Fetch-Site":           {"none"},
-		"Sec-Fetch-User":           {"?1"},
-		"TE":                       {"trailers"},
-		"Update-Insecure-Requests": {"1"},
-		"User-Agent":               {userAgent},
-	}
-
-	response, err := client.Do(request)
-	if err != nil {
-		panic(err)
-	}
-	defer response.Body.Close()
-
-	var result string
-	for _, cookie := range response.Cookies() {
-		if cookie.Name != "AS" {
-			result += cookie.Name + "=" + cookie.Value + "; "
-		}
-	}
-	result = strings.TrimSuffix(result, "; ")
-	return result
-}
-
 // Fetch the latest stock quotes and parse raw fetched data into array of
 // []Stock structs.
 func (quotes *Quotes) Fetch() (self *Quotes) {
@@ -169,9 +89,7 @@ func (quotes *Quotes) Fetch() (self *Quotes) {
 		}
 
 		request.Header = http.Header{
-			"Accept": {"application/json"},
-			// TODO: handle compressed responses
-			// "Accept-Encoding": {"gzip, deflate, br"},
+			"Accept":          {"*/*"},
 			"Accept-Language": {"en-US,en;q=0.5"},
 			"Connection":      {"keep-alive"},
 			"Content-Type":    {"application/json"},
